+++ net_util/Cargo.toml
@@ -11,6 +11,8 @@ log = "0.4.14"
 net_gen = { path = "../net_gen" }
 rate_limiter = { path = "../rate_limiter" }
 serde = "1.0.126"
+versionize = "0.1.6"
+versionize_derive = "0.1.4"
 virtio-bindings = "0.1.0"
 vm-memory = { version = "0.5.0", features = ["backend-mmap", "backend-atomic"] }
 vm-virtio = { path = "../vm-virtio" }
